requirements.txt
psycopg2-binary
fastapi
uvicorn
transformers
pydantic
httpx
sqlalchemy
python-jose[cryptography]
python-dotenv 
passlib[bcrypt] 
alembic 
pytest 
pytest-asyncio

requests
PyJWT
